% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Behavior when attaching the mbg package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{(character(1)) A character string giving the library directory where the
package defining the namespace was found.}

\item{pkgname}{(character(1)) A character string giving the name of the package.}
}
\value{
(invisible) A message may be printed to the console.
}
\description{
Behavior when attaching the mbg package
}
\details{
Yields a message if the INLA package namespace is not available.
}
