% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inla_model.R
\name{fit_inla_model}
\alias{fit_inla_model}
\title{Fit INLA model}
\usage{
fit_inla_model(formula, data_stack, spde, family = "binomial", link = "logit")
}
\arguments{
\item{formula}{(character) INLA formula to fit. Generated in
\code{\link[=prepare_inla_data_stack]{prepare_inla_data_stack()}}, then interpreted using \code{\link[stats:formula]{stats::as.formula()}} within the
call to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{data_stack}{Stacked data, covariates, and spatial index. Generated in
\code{\link[=prepare_inla_data_stack]{prepare_inla_data_stack()}}.}

\item{spde}{SPDE object generated by \code{\link[=prepare_inla_data_stack]{prepare_inla_data_stack()}}.}

\item{family}{(character, default 'binomial') GLM family to use. For more information,
see \code{\link[stats:family]{stats::family()}}.}

\item{link}{(character, default 'logit') Link function to use, typically related to the
GLM \code{family}.}
}
\value{
A fitted INLA model object created by \code{\link[INLA:inla]{INLA::inla()}}
}
\description{
Fit an INLA model based on a constructed data stack and formula
}
\details{
Using \code{\link[INLA:inla]{INLA::inla()}} with reasonable defaults and settings tuned to predict
across a grid. NOTE: The sample size field in the \code{data_stack} MUST be named
"samplesize".
}
