% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging_add_indent}
\alias{logging_add_indent}
\title{Add global indent for logging}
\usage{
logging_add_indent()
}
\description{
Change a global option that controls indentation for MBG package logging
}
\details{
Increases the \code{MbgLoggingIndent} option, with default of one if not set
}
