% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging_drop_indent}
\alias{logging_drop_indent}
\title{Remove global indent for logging}
\usage{
logging_drop_indent()
}
\description{
Change a global option that controls indentation for MBG package logging
}
\details{
Decreases the \code{MbgLoggingIndent} option, with default of zero if not set
}
