% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging_get_timer_log}
\alias{logging_get_timer_log}
\title{Get timer log}
\usage{
logging_get_timer_log(clear_log = FALSE, deindent = TRUE)
}
\arguments{
\item{clear_log}{(\code{logical(1)}, default FALSE) Should the log be cleared afterwards?}

\item{deindent}{(\code{logical(1)}, default TRUE) Should leading whitespace be removed from
timer messages?}
}
\description{
Return a log of all timed events as a data.table
}
\examples{
mbg::logging_start_timer(msg = 'Test logging')
Sys.sleep(0.1)
mbg::logging_stop_timer()
log_results <- mbg::logging_get_timer_log()
print(log_results)

}
