% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_time_stamp}
\alias{make_time_stamp}
\title{Make time stamp}
\usage{
make_time_stamp(suffix = NULL, milliseconds = T)
}
\arguments{
\item{suffix}{(\code{character(1)}, default NULL) suffix to append to the time stamp. Useful
when running batches of related models}

\item{milliseconds}{(\code{logical(1)}, default TRUE) Should milliseconds be appended to
the timestamp? Useful when launching many models in quick succession.}
}
\value{
A string formatted as 'YYYYMMDD_HH_MM_SS\link[=_optional suffix]{_optional MS}'
}
\description{
Create a string time stamp based on current detailed date/time
}
