% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ml_models.R
\name{run_regression_submodels}
\alias{run_regression_submodels}
\title{Run regression sub-models}
\usage{
run_regression_submodels(
  input_data,
  id_raster,
  covariates,
  cv_settings,
  model_settings
)
}
\arguments{
\item{input_data}{A data.frame with at least the following columns:
\itemize{
\item 'indicator': number of "hits' per site, e.g. tested positive for malaria
\item 'samplesize': total population sampled at the site
\item 'x': x position, often longitude
\item 'y': y position, often latitude
}}

\item{id_raster}{terra SpatRaster with non-NA pixels delineating the extent of the
study area}

\item{covariates}{(list) Named list of all covariate effects included in the model,
typically generated by \code{\link[=load_covariates]{load_covariates()}}.}

\item{cv_settings}{Named list of cross-validation settings, passed to
\link[caret:trainControl]{caret::trainControl}.}

\item{model_settings}{Named list where the name of each header corresponds to a model
run in \link[caret:train]{caret::train}, and the arguments correspond to the model-specific settings
for that model type.}
}
\value{
List with two items:
\itemize{
\item "models": A list containing summary objects for each regression model
\item "predictions": Model predictions covering the entire id_raster
}
}
\description{
Wrapper to run many regression sub-models using the caret package
}
