% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_draws.R
\name{summarize_draws}
\alias{summarize_draws}
\title{Summarize draws}
\usage{
summarize_draws(draws, id_fields = NULL, ui_width = 0.95)
}
\arguments{
\item{draws}{A matrix, data.frame, or data.table of predictive draws.}

\item{id_fields}{(default NULL) Only considered for data.frame-like \code{draws}. What
identifier fields in the data should be kept in the summary table and not included
among the draw fields?}

\item{ui_width}{(numeric, default 0.95) Size of the uncertainty interval width when
calculating the upper and lower summary rasters}
}
\value{
A data.table containing at least the following fields:
\itemize{
\item The \code{id_fields}, if passed
\item "mean": Mean across predictive draws
\item "lower": Lower bound of the (X\%) uncertainty interval
\item "upper": Upper bound of the (X\%) uncertainty interval
}
}
\description{
Helper function to summarize a matrix or data.frame of predictive draws
}
