% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging_stop_timer}
\alias{logging_stop_timer}
\title{End logging timer}
\usage{
logging_stop_timer(echo = TRUE)
}
\arguments{
\item{echo}{(\code{logical(1)}, default = TRUE) Should the message be written to screen?}
}
\description{
End a nested timer
}
