% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inla_data_stack.R
\name{prepare_inla_data_stack}
\alias{prepare_inla_data_stack}
\title{Prepare data stack for INLA}
\usage{
prepare_inla_data_stack(input_data, id_raster, covariates)
}
\arguments{
\item{input_data}{A data.frame with at least the following columns:
\itemize{
\item 'indicator': number of "hits' per site, e.g. tested positive for malaria
\item 'samplesize': total population sampled at the site
\item 'x': x position, often longitude
\item 'y': y position, often latitude
}}

\item{id_raster}{terra SpatRaster with non-NA pixels delineating the extent of the
study area}

\item{covariates}{(list) Named list of all covariate effects included in the model,
typically generated by \code{\link[=load_covariates]{load_covariates()}}.}
}
\value{
List containing the following items:
}
\description{
Prepare data stack for INLA
}
